Separate out datatable initialization
