+++ app/javascript/datatables/batchEdit.js
