+++ app/javascript/datatables/index.js
@@ -1,5 +1,5 @@
-import $ from "jquery";
-
+// Load the datatables plugins
+import "jquery";
 import "jszip";
 import "datatables.net-bs4";
 import "datatables.net-buttons-bs4";
@@ -19,14 +19,6 @@ import "datatables.net-buttons-bs4/css/buttons.bootstrap4.css";
 import "datatables.net-rowreorder-bs4/css/rowReorder.bootstrap4.css";
 import "datatables.net-fixedheader-bs4/css/fixedHeader.bootstrap4.css";
 
-import { defaults } from "./config";
-require("./batch_edit");
-
-$(function () {
-  $("table.sortable,table#batch-show").DataTable(defaults);
-  // Bit grim. We register the callback with the legacy jQuery
-  // until we can migrate everything across.
-  window.jQuery(document.body).on("ajaxDomUpdate", function (event, target) {
-    $(target).find("table.sortable").DataTable(defaults);
-  });
-});
+// Load individual datatables implementations
+import "./batchEdit";
+import "./tableSortable";
